LPD-98159 Reuse a shared drop-position helper for rows and groups
